Add test for Utility and External services

+++ src/main/java/com/crowdmap/java/sdk/model/External.java
@@ -127,7 +127,7 @@ public class External implements Serializable {
     /**
      * @param serviceId the serviceId to set
      */
-    public void setServiceId(int serviceId) {
+    public void setServiceId(long serviceId) {
         this.serviceId = serviceId;
     }
